feat: 游戏类型资源订阅搜索

+++ src/components/form/SiteAddEditForm.vue
@@ -1,8 +1,9 @@
 <script lang="ts" setup>
 import { useToast } from 'vue-toast-notification'
+import { useConfirm } from 'vuetify-use-dialog'
 import type { Site } from '@/api/types'
 import { doneNProgress, startNProgress } from '@/api/nprogress'
-import { numberValidator, requiredValidator } from '@/@validators'
+import { numberValidator, requiredValidator, urlPathValidator, urlValidator } from '@/@validators'
 import api from '@/api'
 
 // 输入参数
@@ -18,7 +19,20 @@ const emit = defineEmits(['save', 'remove', 'close'])
 const siteForm = ref<Site>({
   id: props.siteid ?? 0,
   url: '',
-  rss: '',
+  types: [],
+  feed: {
+    method: 'GET',
+    path: '',
+    body: '',
+    rule: '',
+  },
+  search: {
+    method: 'GET',
+    path: '',
+    body: '',
+    rule: '',
+  },
+  xpath: '',
   cookie: '',
   ua: '',
   pri: 0,
@@ -31,6 +45,9 @@ const siteForm = ref<Site>({
 
 // 提示框
 const $toast = useToast()
+
+// 确认框
+const deleteConfirm = useConfirm()
 
 // 状态下拉项
 const statusItems = [
@@ -45,6 +62,26 @@ const priorityItems = ref(
     value: item,
   })),
 )
+
+// 请求方式下拉框选项
+const httpMethodItems = [
+  { title: 'HTTP GET', value: 'GET' },
+  { title: 'HTTP POST', value: 'POST' },
+]
+
+// 订阅方式下拉框选项
+const feedMethodItems = [
+  { title: 'RSS', value: 'RSS' },
+  { title: 'HTTP GET', value: 'GET' },
+]
+
+// 适用类型下拉框选项
+const mediaTypeItems = [
+  { title: '电影', value: 'MOVIE' },
+  { title: '电视剧', value: 'TV' },
+  { title: '动漫', value: 'ANIME' },
+  { title: '游戏', value: 'GAME' },
+]
 
 // 监控输入参数
 watchEffect(async () => {
@@ -86,6 +123,22 @@ async function addSite() {
 
 // 调用API删除站点信息
 async function deleteSiteInfo() {
+  const isConfirmed = await deleteConfirm({
+    title: '确认',
+    content: `是否确认删除【${siteForm.value?.name}】?`,
+    confirmationText: '确认',
+    cancellationText: '取消',
+    dialogProps: {
+      maxWidth: '50rem',
+    },
+    confirmationButtonProps: {
+      variant: 'tonal',
+    },
+  })
+
+  if (!isConfirmed)
+    return
+
   try {
     const result: { [key: string]: any } = await api.delete(`site/${siteForm.value?.id}`)
     if (result.success)
@@ -136,13 +189,23 @@ async function updateSiteInfo() {
               md="6"
             >
               <VTextField
-                v-model="siteForm.url"
-                label="站点地址"
+                v-model="siteForm.name"
+                label="站点名称"
                 :rules="[requiredValidator]"
               />
             </VCol>
             <VCol
-              cols="12"
+              cols="6"
+              md="3"
+            >
+              <VSelect
+                v-model="siteForm.is_active"
+                :items="statusItems"
+                label="状态"
+              />
+            </VCol>
+            <VCol
+              cols="6"
               md="3"
             >
               <VSelect
@@ -154,22 +217,74 @@ async function updateSiteInfo() {
             </VCol>
             <VCol
               cols="12"
-              md="3"
+              md="6"
+            >
+              <VTextField
+                v-model="siteForm.url"
+                label="站点地址"
+                :rules="[requiredValidator, urlValidator]"
+              />
+            </VCol>
+            <VCol
+              cols="12"
+              md="6"
             >
               <VSelect
-                v-model="siteForm.is_active"
-                :items="statusItems"
-                label="状态"
+                v-model="siteForm.types"
+                label="适用类型"
+                multiple
+                :items="mediaTypeItems"
+                :rules="[requiredValidator]"
+              />
+            </VCol>
+          </VRow>
+          <VRow>
+            <VCol v-if="siteForm.feed" cols="3">
+              <VSelect
+                v-model="siteForm.feed.method"
+                label="FEED方式"
+                :items="feedMethodItems"
+              />
+            </VCol>
+            <VCol v-if="siteForm.feed" cols="9">
+              <VTextField
+                v-model="siteForm.feed.path"
+                label="FEED地址路径，页码用'{page}'代替"
+                :rules="[urlPathValidator]"
+              />
+            </VCol>
+          </VRow>
+          <VRow>
+            <VCol v-if="siteForm.search" cols="3">
+              <VSelect
+                v-model="siteForm.search.method"
+                label="SEARCH方式"
+                :items="httpMethodItems"
+              />
+            </VCol>
+            <VCol v-if="siteForm.search" cols="9">
+              <VTextField
+                v-model="siteForm.search.path"
+                label="SEARCH地址路径，搜索关键字用'{keyword}'代替"
+                :rules="[urlPathValidator]"
+              />
+            </VCol>
+            <VCol v-if="siteForm.search && siteForm.search.method === 'POST'" cols="12">
+              <VTextarea
+                v-model="siteForm.search.body"
+                label="SEARCH请求体"
               />
             </VCol>
           </VRow>
           <VRow>
             <VCol cols="12">
-              <VTextField
-                v-model="siteForm.rss"
-                label="RSS地址"
+              <VTextarea
+                v-model="siteForm.xpath"
+                label="XPATH规则"
               />
             </VCol>
+          </VRow>
+          <VRow>
             <VCol cols="12">
               <VTextarea
                 v-model="siteForm.cookie"
